Add files via upload
Query 1: 
o Was the average of global sales higher before or after 2010 ?

The only way to make this query work is because of using the CASE STATEMENT as this will enable us to have a separate column that will mark the result after using the IF,ELSE clause within the CASE STATEMENT.
In my code I have used the comparison of newly created columns that were named as before_2010 and after_2010 by using WHERE Year<2010 and Year> 2010 respectively and these 2 columns will indicate the Average Global_Sales values. These 2 columns are named as a and b respectively.
Now my CASE STATEMENT will make use of comparison between the values of these two global sales which I have created for years before and after 2010. Based on this comparison the result will be displayed in a separate column which is mentioned in closing of CASE STATEMENT as ‘Solution’.

Query type 2:

o	Create a new column that labels records before 2010 as 'pre-2010' and after 2010 as 'post-2010'

In this query I have returned Name of the game ,platform and a new column named as record. This new column was created using the CASE STATEMENT where I have provided the constraint for year value of before and after 2010. Before 2010 is stated as pre-2010 and after 2010 the record column will show post-2010 as the returning value for all the games in the vgsales.
I have also used ORDER BY statement with respect to NAME in ascending order to get the list in alphabetical order.